refactor(stations): 노선 배지 presentation 경계 분리
